Added Game Log so players can see what happened.
Moved voting logic to server side. Players no longer click "Ready" to
record their vote, it is recorded immediately upon choosing. Players can
change vote as long as phase has not ended.
Removed timeouts for showing the day
phase, we can make this a popup instead. As soon as vote is unanimous
for villains or majority for day phase the game progresses to next
phase. (Still need to add the extra night phases for guardian and
telepath). Game log will make it so people can see what happened in
previous phase. We can also add a popup to show who was killed that
players can close.