Add guardian log, shows only on guardian's screen.
They can thus keep track of their history as well as the main game
history. Still need to: 1) Figure out what happens when Guardian is dead
during the night. rest of game must not know that guardian is dead. And
need to implement rule that you cant save same person twice in a row.
And you *are* supposed to be able to save yourself as guardian (just not twice in a
row).is should help them know what they did the last time.